move `useSiteAnalysis` and dependants from `data-stores` to `help-center` package
