+++ packages/data-stores/src/index.ts
@@ -13,9 +13,6 @@ import * as User from './user';
 import * as WpcomPlansUI from './wpcom-plans-ui';
 export { useHappinessEngineersQuery } from './queries/use-happiness-engineers-query';
 export { useJetpackSearchAIQuery } from './queries/use-jetpack-search-ai';
-export { useSiteAnalysis } from './queries/use-site-analysis';
-export { useUserSites } from './queries/use-user-sites';
-export type { AnalysisReport } from './queries/use-site-analysis';
 export { useSiteIntent } from './queries/use-site-intent';
 export * from './starter-designs-queries';
 export * from './site/types';
