+++ packages/help-center/src/components/help-center-notice.tsx
@@ -4,7 +4,7 @@ import { createInterpolateElement } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 import { Icon, info } from '@wordpress/icons';
 import { SupportActivity } from '../data/types';
-import type { AnalysisReport } from '@automattic/data-stores';
+import type { AnalysisReport } from '../data/types';
 import type { ReactNode } from 'react';
 
 type Props = {
